Transformation to ODE System

This section provides functions to transform DAE to ODE systems. In particular,

to find equations that need to be differentiated, in order to transform a Differential Algebraic Equations (DAEs) agebraically to Ordinary Differential Equations (ODEs),

to differentiate relevant equations analytically,

to statically select ODE states and transform to ODE form.

Main Functions

ModiaBase.BLTandPantelides.pantelides! — Function

function pantelides!(G, M, A)

Perform index reduction with Pantelides algorithm.

G: bipartite graph (updated)

M: number of V-nodes

A: A[j] = if V[k] = der(V[j]) then k else 0

return assign: assign[j] contains the E-node to which V-node j is assigned or 0 if V-node j not assigned

return A: A[j] = if V[k] = der(V[j]) then k else 0

return B: B[i] = if E[l] = der(E[i]) then l else 0

Reference: Pantelides, C.: The consistent initialization of differential-algebraic systems. SIAM Journal of Scientific and Statistical Computing, 9(2), pp. 213–231 (1988).

ModiaBase.Differentiate.derivative — Function

der = derivative(ex, timeInvariants=[])

Form the derivative of the expressions ex with regards to time. Time derivative of variables are denoted der(v).

ex: Expression to differentiate

timeInvariants: Vector of time invariant variables, i.e. with time derivative equal to zero.

returnder`: Time derivative of ex

ModiaBase.StateSelectionFunctions — Type

stateSelectionFunctions = StateSelection.StateSelectionFunctions(;
    var_name               = v          -> nothing,
    var_julia_name         = v          -> nothing,
    var_unit               = v          -> "",
    var_startInitFixed     = v          -> (nothing,false),   # (startOrInit, fixed)
    var_is_state           = v_original -> false,
    equation               = e          -> ""
    isSolvableEquation     = (e_original,v_original) -> false,
    isLinearEquation       = (e_original,v_original) -> (false, false),
    getSolvedEquationAST   = (e, v)                  -> nothing,
    getResidualEquationAST = (e, residualName)       -> nothing,
    showMessage            = (message; severity=0,from="???",details="",variables=Int[],equations=Int[]) -> println("...")
)

Generate an instance of an immutable struct that holds the callback functions as needed by getSortedAndSolvedAST.

The functions need to have the following arguments:

var_name(v::Int)::String:
Return full (Modia) path name of v as String (e.g. "a.b.c" or "a.b.der(c)").

var_julia_name(v::Int)::String:
Return full Julia path name of variable v as Symbol (e.g. Symbol("m.a.b.c") or Symbol("abder_c")).

var_unit(v::Int)::String:
Return unit of variable v as String (for example "N*m") or "", if no unit is defined.

var_startInitFixed(v::Int)::Tuple(Any,Bool):
Return (startOrInit, fixed), where startOrInit is the start or init value or nothing (if neither start nor init defined) and fixed is true, if an init value is defined.

var_is_state(v_original::Int)::Bool:
Return true, if variable v_original is defined to be a state.

equation(e::Int):
Return equation as string or "", if equation is not provided.

isSolvableEquation(e_original::Int, v_original::Int)::Bool:
Return true if original, undifferentiated equation e_original can be uniquely solved for variable v_original.

(isLinear, hasConstantCoefficient) = isLinearEquation(e_original::Int, v_original::Int):
Return the information whether the original, undifferentiated equation e is linear with respect to variable v (isLinear=true) and if this is the case whether the coefficient of this linear term is constant after initialization (hasConstantCoefficient=true). If v is an array, then the array equation e must be linear with respect to all elements of array v in order to return isLinear=true.

AST = getSolvedEquationAST(e:Int, v:Int):
Return Abstract Syntax Tree AST::Expr of equation e solved for variable v as instance of Expr.

AST = getResidualEquationAST(e::Int, residualName::Symbol):
Return Abstract Syntax Tree AST::Expr of eqation e as residual equation with ustrip(..)
residualName = ustrip( < residual form of equation e> )

showMessage(message; severity=0,from="???",details="",variables=Int[],equations=Int[]):
Print information, warning or error message.

ModiaBase.getSortedAndSolvedAST — Function

(success, AST, equationInfo) = getSortedAndSolvedAST(
                                    G, BLT, assign, A, B, stateSelectionFunctions;
                                    log = false, logDetails=false, modelName = "???",
                                    defaultParameterAndStartValues = nothing,
                                    unitless=false)

From the BLT on highest derivative level (G, BLT, assign, A, B), a set of callback functions stateSelectionFunctions, this function returns the AST (Abstract Syntax Tree) of the sorted and solved equations as a vector of Expr and information about the ODE equations (especially names of the ODE states).

Input arguments

G: G[i] is the Integer vector of variable indices of equation i. A variable i can be a scalar or an array. is of type Vector{ Vector{Int} } or Vector{ Any }.

BLT::Vector{Vector{Int}}: BLT[i] is the vector of equations belonging to BLT-block i. In principal, BLT[i] must include only highest derivative equations. In order to ease the generation of BLT[i], to not be forced to map indices forth and back when calling the function that computes the BLT, the following assumption is made: If BLT[j] contains only one equation with B[j][1] > 0 then this BLT block is a lower-derivative (dummy) block and is ignored.

assign::Vector{Int}: ei = assign[vi] is equation ei assigned to variable vi

A: A-Vector of Pantelides algorithm: A[i] = if der(v[i]) == v[k] then k else 0 where v[i] is variable i.

B: B-Vector of Pantelides algorithm: B[i] = if der(e[i]) == e[k] then k else 0 where e[i] is equation i.

stateSelectionFunctions: An instance of immutable struct StateSelectionFunctions that holds the information about the callback functions needed by getSortedAndSolvedAST.

log: = true: Print debug information (to find bugs in the code).

logDetails: = true: Print detailed debug information, if log = true.

modelName: Name of model used in messages (otherwise not used).

Output arguments

A tuple with the following values:

success::Bool: = true, if generation was successful. = false, if an error occured. Arguments AST, equationInfo hold the information that have been collected until the error was detected.

AST::Vector{Expr}: The AST (Abstract Syntax Tree) of the sorted and solved equations as Vector of Expr.

equationInfo::ModiaBase.EquationInfo: Object that defines the states of the ODE.

When variable v has neither a start nor init attribute defined, it is assumed to have start=0.0.

defined	Description

start	start is allowed to be changed during initialization

init	init is not allowed to be changed during initialization

Note, start, init influence the state selection.

If an ODE state has neither a start nor an init value defined, a warning message is printed (start/init value missing).

If a solved variable has an init value, an information message is printed in case log=true (init value has no effect).

Sketch of the Algorithm

Note, a statement error xxx means that an error occurs, because it is not possible to transform the equations to an ODE.

Generate equation sets

From BLT blocks that are not further differentiated (= highest order BLT blocks), generate the equation sets as described by the dummy derivative method of Mattsson and Söderlind (1992). See also sections 4.5 and 4.6 of the paper Otter and Elmqvist (2017).

Define initial ODE states

Initially, all differentiated variables are defined to be ODE states. Whenever an initial ODE state is explicitly solved for in the course of the algorithm, the variable is no longer an ODE state.

Analyse equation sets

For every equation set on every differentiation level perform the following actions:

If the equation set consists of one equation in one unknown, the equation is solved for this unknown (error, if this is not possible).

If the equation set consists of N linear equations in N unknowns solve this equation system (error, if no linear system):
The equations are first teared to reduce the number of iteration variables and afterwards the teared equation system is solved with a special iterator loop that solves a linear equation system with an LU decomposition with column pivoting (for details see ModiaBase.LinearEquationsIteration!).

If the equation set consists of N linear equations in M unknowns (M > N) perform the following actions (error, if no linear system).
Note, due to the structure of BLT, this equation set cannot be on highest derivative level; due to the Pantelides algorithm, M < N is not possible).

Solve the equation system explicitly via tearing for N unknowns (error, if this is not possible). Use start/init attributes of the M unknowns to influence tearing (e.g. it is tried to eliminate first variables that have neither start not init values).

The N solved variables have been initially ODE states and are now defined to be no ODE states (so called dummy states, that is algebraic variables).

A BLT transformation of all equations is made under the assumption that the selected ODE states are known. This phase is needed, because the dummy-derivative ordering at the beginning does not necessarily provide already the right ordering. In this phase the information from the previous phase is used (that is the already determined tearing variables for systems of equations are utilized).

Utility Functions

ModiaBase.StateCategory — Type

@enum StateCategory

Type of a state.

value	description

XD	Differential variable (derivative of variable appears in equations)

XALG	Algebraic variable that is included in x

XLAMBDA	Algebraic variable that is included in der_x

XMUE	Algebraic variable used for stabilization and included in der_x (exact value = 0)

ModiaBase.ResidualCategory — Type

@enum ResidualCategory

Type of a residual (see also StateCategory for the type of a state).

value	description

FD	Differential equation (depends on XD, XLAMBDA, XMUE variables)

FC_ALG	Constraint equation that is purely algebraic (depends on XALG variables)

FC_LOW_HIGH	Constraint equation on lowest and on highest derivative level

FC_MUE	Constraint equation that is associated with a stabilizer XMUE

ModiaBase.LinearEquations — Type

leq = LinearEquations{FloatType}(x_names::Vector{String}, x_vec_julia_names::AbstractVector,
                                 x_lengths::Vector{Int}, nx_fixedLength::Int, A_is_constant::Bool;
                                 useRecursiveFactorizationUptoSize = 0)

Define linear equation system "A*x=b" with x::Vector{FloatType}.

x is constructed from a set of scalar or vector variables i with names x_names[i] and length x_length[i] (that is length(x) = sum(x_lengths)). xnames[1:nxfixedLength] are elements with fixed lengths (dimensions do not change after compilation). xnames[nxfixedLength+1:end] are vector-valued elements where the dimensions may changer after compilation.

x_vec_julia_names are the Julia names of the vector-valued elements.

If A_is_constant = true then A is a matrix that is constant after initialization

If length(x) <= useRecursiveFactorizationUptoSize, then linear equation systems will be solved with RecursiveFactorization.jl instead of the default lu!(..) and ldiv!(..).

For details how to use this constructor, see LinearEquationsIteration!.

ModiaBase.LinearEquationsIteration! — Function

 iterating = LinearEquationsIteration!(leq, isInitial, [solve, isStoreResult,] time, timer)

This function solves a linear equation system in residual form "residual = A*x - b" by iterating with a while loop over this system (arguments solve, isStoreResult are optional and have only an effect if leq.odeMode=false, that is if the linear equation system is solved in DAE mode):

function getDerivatives!(_der_x, _x, _m, _time)::Nothing
    _leq::Union{Nothing,LinearEquations{FloatType}} = nothing
    ...
    _leq      = _m.linearEquations[<nr>]   # leq::LinearEquations{FloatType}
    _leq.mode = -3  # initializes the iteration
    while LinearEquationsIteration!(_leq, _m.isInitial, _m.solve_leq, _m.storeResult, _m.time, _m.timer)
        x1 = _leq.x[1]
        x2 = _leq.x[2]
        x3 = _leq.x_vec[1]
        ...
        v_solved1 = f(x1, x2, ..., positive(x1,.., _leq))
        v_solved2 = f(x1, x2, ..., v_solved1)
        ...
        appendResidual!(_leq.residuals, < getResiduals(x1,x2,...) > )
    end
    ...
end

Note, A and b can be functions of linear event operators, such as positive(c_i'*x - d_i). In this case this system is solved by a fixed point iteration scheme, that is A*x = b is solved until potentially present positive(c_i'*x - d_i) calls are consistent to x (c_i'*x - d_i and positive(..) must have the same sign). Iteration is terminated after 20 iterations if no convergence is found. In this case a warning message is triggered and simulation is continued (it might be that simulation is still successful, even if x and positive(..) are temporarily not consistent to each other).

The current values of A,x,b,residuals are stored in leq. If A is fixed after initialization (leq.Aisconstant = true), then A is computed only once at initialization, the LU decomposition of A is stored in leq and used in subsequent calls to solve the equation system.

Input arguments

leq::LinearEquations{FloatType}: Instance of LinearEquations.

isInitial::Bool: = true: Called during initialization.

solve::Bool: = true: leq.x is computed by LinearEquationsIteration!. = false: leq.x has been set by calling environment (typically when called from DAE integrator). Note, at events and at terminate, solve must be true).

isStoreResult::Bool: = true: Called at a communication point (store results).

time: Actual time instant (only used for warning/error messages).

timer::TimerOutputs.TimerOutput: Timer used to measure the solution of the linear equation system.

Output arguments

iterating::Bool: = true : Continue iteration of while-loop. = false: Immediately break while-loop after this function was terminated.

Enhancing efficiency

Variable leq.mode::Int can be accessed (read-only) in the body of the while-loop to enhance efficiency. For example, if residuals are computed in a function and the function evaluation is expensive.

leq.mode = -2: Residuals might be computed, but they are not used.
leq.mode = -1: Compute "residuals .= A*x - b"
leq.mode =  0: Compute "residuals .= A*0 - b"
leq.mode >  0: Compute "residuals .= A*e_i - b"   # e_i = unit vector with i = leq.mode

Hidden argument leq.mode::Int on input

leq.mode = -3  # LinearEquationsIteration! is called the first time
               if leq.odeMode || solve
                  # ODE mode or DAE mode at an event (solve "x" from equation "residuals = A*x - b")
                  # Initialize fixed point iteration or continue fixed point iteration (if in DAE mode)
                  leq.niter     = 0      # Number of event iterations
                  leq.success   = false  # Event iteration was not yet successful
                  leq.x        .= 0
                  leq.mode      = 0      # Compute "residuals .= A*0 - b"
               else # solve = false
                  # DAE mode (but not at an event)
                  leq.niter     = 0
                  leq.success   = true
                  leq.x_is_zero = false
                  if isStoreResult
                     leq.mode = -2  # Residuals might be computed, but they are not used.
                  else
                     leq.mode = -1  # Compute "residuals .= A*x - b"
                  end
               end
               return true  # Continue while-loop

leq.mode = -2  # Terminate while-loop or initialize next event iteration
               if leq.success           # Set from leq.mode=-2 or leq.mode>=0 or from event operators, such as positive!(...)
                  return false          # Terminate while-loop
               elseif leq.niter > 20
                  <warning>             # Event iteration failed (variables are not consistent)
                  return false          # Terminate while-loop
               else
                  # Initialize next event iteration
                  leq.x        .= 0
                  leq.x_is_zero = true
                  leq.mode      = 0     # Compute "residuals .= A*0 - b"
                  return true           # Continue while-loop
               end

leq.mode = -1: @assert(!leq.odeMode && !solve)
               # DAE mode, but not at an event
               return false      # Terminate while-loop

leq.mode =  0: @assert(leq.odeMode || solve)
               # ODE mode or DAE mode at an event (solve "x" from equation "residuals = A*x - b")
               leq.b    = -leq.residuals
               leq.x[1] = 1.0
               leq.mode = 1      # Compute "residuals := A*e_1 - b"
               return true       # Continue while-loop

leq.mode >  0: @assert(leq.odeMode || solve)
               j          = leq.mode
               leq.A[:,j] = leq.residuals + leq.b
               leq.x[j]   = 0.0
               if j < nx
                   leq.x[j+1] = 1.0
                   leq.mode   = j+1      # Compute A[:,j+1] after the next iteration
                   return true           # Continue while-loop
               elseif j == nx
                   leq.x = solve(leq.A,leq.b) # Solve linear equation system A*x = b for x.
                   leq.mode    = -1      # Compute all variables IN the next iteration as function of x
                   leq.niter  += 1       # Increment number of iterations to solve A*x = b
                   leq.success = true    # Terminate for-loop at the beginning of the next iteration,
                                         # provided no positive(..) call changes its value.
                                         # (leq.success is set to false in positive(..), if the return value changes).
                   return true           # Continue while-loop
               end

ModiaBase.EquationInfo — Type

eqInfo = EquationInfo(;
            status               = MANUAL,
            ode                  = true,
            nz                   = 0,
            x_info               = StateElementInfo[],
            nx                   = -1,
            nxFixedLength        = nx,
            residualCategories   = ResidualCategory[],
            linearEquations      = Tuple{Vector{String},AbstractVector,Vector{Int},Int,Bool}[],
            vSolvedWithFixedTrue = String[],
            defaultParameterAndStartValues = nothing,
            ResultType = nothing,
            ResultTypeHasFloatType = false)

Return instance eqInfo that defines the information for the equation system.

Arguments

status::EquationInfoStatus: Defines the variables that have a value.

ode: = true if ODE-interface (getDerivatives!), = false if DAE-interface (getResiduals!).

nz: Number of zero crossing functions.

x_info: Vector of StateElementInfo elements provding info for every x-element

residualCategories: If ode=true, length(residualCategories) = 0. If ode=false: residualCategories[i] is the ResidualCategory](@ref) of x-element "i".

linearEquations::Vector{Tuple{Vector{String},AbstractVector,Vector{Int},Int,Bool}}: linearEquations[i] defines a ModiaBase.LinearEquations system, where the first tuple value is a vector of the names of the unknowns, the second tuple value is a vector of the Julia names of the vector-valued elements, the third tuple value is a vector with the lengths of the unknowns, the fourth tuple value is the number of residuals and the fifth tuple value defines whether the coefficient matrix A has only constant entries (=true) or not (=false).

vSolvedWithFixedTrue::Vector{String}: Vector of variables that are computed from other variables and have fixed=true. During initialization it is checked whether the calculated values and the start values of these variables agree. If this is not the case, an error is triggered.

defaultParameterAndStartValues: Dictionary of default paramter and default start values.

ResultType::AbstractModelValues: ModelValues type that shall be used for the result generation (= ModelValues struct without parameters).

ResultTypeHasFloatType::Bool: = false, if ResultType is not parameterized. = true, if ResultType has parameter FloatType.

ModiaBase.StateElementInfo — Type

xe_info = StateElementInfo(...)

Return an instance of the mutable struct StateElementInfo that defines the information for one element of the state vector. There are three constructors:

Default constructor (all variables under section Arguments are given; used to write/read the complete information).

All variables under section Arguments are given with exception of startIndex (used for EquationInfoStatus = CODE_GENERATION).

StateElementInfo(xname, derx_name, stateCategory; fixed=nothing, nominal=NaN, unbounded=false) (used for EquationInfoStatus = MANUAL):

Arguments

x_name: Name of x-element or "" if no name (if stateCatebory = XLAMBDA or XMUE)

xnamejulia: Julia name of x-element in getDerivatives!/getResiduals! function or "" if not needed (since no code generation).

derxname: Name of derx-element or "" if either `der(xname)` or if no name (if stateCategory = XALG).

derxnamejulia: Julia name of derx-element in getDerivatives!/getResiduals! function or "" if not needed (since no code generation).

stateCategory::StateCategory: Category of the state

length: length of x-element (or -1 if not yet known)

unit: unit of XD, XALG (xname) or XLAMBDA, XMUE (derx_name) or "" if not yet known)

fixed: false (= guess value) or true (= not changed by initialization). Only relevant for ode=false, otherwise ignored.

nominal: Nominal value (NaN if determined via start value)

unbounded: false or true

startIndex: start index of x-element with respect to x-vector or -1 if not yet known.